\name{as.binary-methods}
\docType{methods}
\alias{as.binary-methods}
\alias{as.binary,VegsoupData-method}
\alias{as.character-methods}
\alias{as.character,VegsoupData-method}
\alias{as.numeric-methods}
\alias{as.numeric,VegsoupData-method}
\title{Methods to Cast Species Matrix of Objects Extending \dQuote{VegsoupData}}
\description{
The function creates a matrix of mode according the methods name. 
}
\section{Methods}{
\describe{
See \code{\link{as.binary}}, \code{\link{as.numeric}}, or \code{\link{as.character}} for more details.
}
}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
