\name{as.logical-methods}
\docType{methods}
\alias{as.logical-methods}
\alias{as.logical,VegsoupData-method}
\alias{as.character-methods}
\alias{as.character,VegsoupData-method}
\alias{as.numeric-methods}
\alias{as.numeric,VegsoupData-method}
\title{Methods to Cast Species Matrix of Objects Extending \dQuote{VegsoupData}}
\description{
The function creates a matrix of mode according the methods name. 
}
\section{Methods}{
\describe{
See \code{\link{as.logical}}, \code{\link{as.numeric}}, \code{\link{as.character}},
or \code{\link{as.matrix}} for more details.
}
}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
