\name{dim-methods}
\docType{methods}
\alias{dim-methods}
\alias{dim,VegsoupData-method}
\title{Methods for Function dim}
\description{
Dimensions of an Object Extending Class VegsoupData
}
\section{Methods}{
\describe{
dim is a method for the \code{slot(obj, "species")} an retrieves the dimension of the species data set disregarding occurences of a species on different layers. Rows are plots ans columns are speciesSo \code{dim(obj)[2]} reports the 
\item{\code{signature(x = "VegsoupData")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
