\name{getK}
\docType{methods}
\alias{getK}
\alias{getK-methods}

\title{ Return the number of Partitions of an Object }
\description{
  Simple getter method and returns the number of partitions (cluster)
}
\usage{
  getK(obj)
}
\arguments{
  \item{obj}{
    An pbject inheriting from class \code{"VegsoupPartition"}.
  }
}
\value{
  A single integer.
}
\author{ Roland Kaiser }
\seealso{
  \code{\linkS4class{VegsoupPartition}}
}
\keyword{ methods }

