\encoding{UTF-8}

\name{hist}

\alias{hist}
\alias{hist,Vegsoup-method}
\alias{hist,VegsoupPartition-method}
\alias{hist,VegsoupPartitionFidelity-method}

\title{ Plot Histogram of Fidelity Values }
\description{
  Plots a histogram of fidelity values.
}
\usage{
  hist(x, ...)
}
\arguments{
  \item{x}{
    \code{'VegsoupPartitionFidelity'} object.
  }
  \item{\dots}{
    arguments passed to \code{\link{hist}}
  }
}
\details{
  See to \code{\link{hist}} for details.
}
\value{
  An object of class \code{"histogram"}.
}
\author{ Roland Kaiser }
\keyword{ methods }

