\name{Multipatt-methods}
\docType{methods}

\alias{Multipatt-methods}
\alias{Multipatt,VegsoupPartition-method}

\title{ Methods for Function Multipatt }
\description{
 Nothing here yet.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "VegsoupPartition")}}{
 Nothing here yet.
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
