\name{names-methods}
\docType{methods}
\alias{names-methods}
\alias{names,VegsoupData-method}
\title{Methods for Function names}
\description{
%all.equal(dimnames(as.binary(dta))[[2]], names(dta))
}
\section{Methods}{
\describe{

\item{\code{signature(x = "VegsoupData")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
