\name{foo}
\alias{foo}
\title{foo}

\description{
Creates a "Google Maps" stand-alone web page using Google Visualisation API supplied from package googleVis in function.
}
\usage{QuickMap(obj)}

\details{
See \link{gvisMap} in package \code{googleVis} for details.
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
}
\keyword{datasets}
