\name{testdata}

\alias{testdata}
\alias{spc}
\alias{sts}
\alias{txn}
\alias{barmstein}

\docType{data}

\title{ Test Data Set }

\description{
Simple data structures for testing. An object created by \code{\link{coenoflex}} with 100 plots and 200 species.
}

\format{
Vegsoup object.
}

\usage{
data(barmstein)
data(testdata)
}

\examples{
data(testdata)
}
