\name{testsites}
\alias{testsites}
\docType{data}
\title{Test data set for sites}
\description{
	Simple data structure for testing.
}
\format{
This data frame contains the following columns:
\describe{
	\item{plot}{plot identifier, mode \code{"character"}}
	\item{variable}{name of the variable, mode \code{"character"}}
	\item{value}{value for each variable, mode \code{"character"}}
}
}
\usage{
data(testsites)
}	
\examples{
data(testsites)
}