\name{Spread-methods}
\docType{methods}
\alias{Spread-methods}
\alias{Spread,VegsoupDataPartition-method}
\title{Methods for Function Spread}
\description{
\code{Spread} returns a list of species occurences and cluster assignments.
}
\section{Methods}{
\describe{

\item{\code{signature(obj = "VegsoupDataPartition")}}{

}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
