\name{testspecies}
\alias{testspecies}
\docType{data}
\title{Test data set for species}
\description{
	Simple data structure for testing.
}
\format{
This data frame contains the following columns:
\describe{
	\item{plot}{plot identifier, mode \code{"character"}}
	\item{abbr}{species names, mode \code{"character"}}
	\item{layer}{layer code, mode \code{"character"}}
	\item{cov}{species abundance, mode \code{"character"}} 	
}
}
\usage{
data(testspecies)
}	
\examples{
data(testspecies)
}