\name{testtaxonomy}
\alias{testtaxonomy}
\docType{data}
\title{Test data set for taxonomy}
\description{
	Simple data structure for testing.
}
\format{
This data frame contains the following columns:
\describe{
	\item{abbr}{species identifier, mode \code{"character"}}
	\item{taxon}{scientific species name, mode \code{"character"}}
}
}
\usage{
data(testspecies)
}	
\examples{
data(testspecies)
}