\name{vegsoup-package}
\alias{vegsoup-package}
\alias{vegsoup}
\docType{package}
\title{ An overview of the functions in this package }
\description{
The vegsoup package provides classes and functions to manipulate vegetation data. It can be seen as a data base system. Vegetation data consists at least of two related tables -- a species matrix and a table of sites properties, including measured environmental varibales. If taxonomic idenity, or traits of species is a concern an additional table has to be tightly linked the species matrix. Moreover, if spatial representation (e.g. plot coordinates and/or polygons) needs to be adressed another data layer has to be included. In the most complete case we end up with indipendent R objects. Indexing and manipulation easily get cumbersome in this situation and the general workflow is hindered by a serious amount code that ensures linking of objects 

The package should be particularly usefull when \dots.

Below is a list of functions grouped by theme.
}
\details{
\tabular{ll}{
Package: \tab vegsoup\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2011-06-27\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
Depends: \tab methods\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Roland Kaiser

Maintainer: kardinal.eros@gmail.com

}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
