\name{getStat}
\docType{methods}

\alias{getStat}
\alias{getStat,Vegsoup-method}

\title{ Methods for \code{VegsoupPartitionFidelity} Objects }
\description{
  Return fidelity statistics.
}
\usage{
  getStat(obj, ...)
}
\arguments{
  \item{obj}{
    A \code{"VegsoupPartitionFidelity"} object.
  }
  \item{\dots}{
    Not used.
  }
}
\value{
  A \code{"matrix"} of fidelity (indicator) values.
}
\author{ Roland Kaiser }
\keyword{ methods }
