\name{Partitioning}
\docType{methods}
\alias{Partitioning}
\alias{Partitioning-methods}

\title{ Getter Methods for \code{"VegsoupPartition"} Objects}
\description{
  Retreive class slots.
}
\usage{
Partitioning(obj)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obj}{
    An object inheriting from class \code{\linkS4class{VegsoupPartition}}.
}
}

\value{
  Returns the running partitioning vector.
}
\author{ Roladn Kaiser }
\seealso{
\code{\linkS4class{getK}}, ~~~
}
\keyword{ methods }

