\name{[-methods}
\docType{methods}
\alias{[-methods}
\alias{[,VegsoupData,ANY,ANY,missing-method}
\alias{[,VegsoupDataPartition,ANY,ANY,missing-method}
\title{Methods for Function [}
\description{
obj[c(1,1,2), ] is currently not implemented
}
\section{Methods}{
\describe{

\item{\code{signature(x = "VegsoupData", i = "ANY", j = "ANY", drop = "missing")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(x = "VegsoupDataPartition", i = "ANY", j = "ANY", drop = "missing")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{ ~~ other possible keyword(s) ~~ }
